requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
ffmpeg-python==0.2.0
boto3==1.29.7
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml
pydantic-settings
apscheduler
moviepy==1.0.3
python-dotenv
aiofiles

django-scheduler
pyyaml
cronex
pydantic-settings